PEP8

+++ python-sdk/nuscenes/eval/detection/algo.py
@@ -176,7 +176,7 @@ def calc_ap(md: MetricData, min_recall: float, min_precision: float) -> float:
     prec = prec[round(100 * min_recall) + 1:]  # Clip low recalls.
     prec -= min_precision  # Clip low precision
     prec[prec < 0] = 0
-    return np.mean(prec) / (1.0 - min_precision)
+    return float(np.mean(prec)) / (1.0 - min_precision)
 
 
 def calc_tp(md: MetricData, min_recall: float, metric_name: str) -> float:
